requirements.txt
aiohttp==3.9.5
botbuilder-core==4.17.0
botbuilder-dialogs==4.17.0
botbuilder-schema==4.17.0
azure-identity==1.14.0
azure-ai-language-conversations==1.1.0b1
pyodbc==4.0.39
gunicorn==22.0.0
pytest
